Add connection resolution
