+++ apps/monolith/src/ContosoUniversity.Data/IConnectionResolver.cs
@@ -1,0 +1,8 @@
+namespace ContosoUniversity.Data;
+
+using Microsoft.Data.SqlClient;
+
+public interface IConnectionResolver
+{
+    SqlConnectionStringBuilder CreateFor(string connectionStringName);
+}
