+++ apps/monolith/src/ContosoUniversity.Data/StartupExtensions.cs
@@ -1,0 +1,13 @@
+namespace ContosoUniversity.Data;
+
+using Connection;
+
+using Microsoft.Extensions.DependencyInjection;
+
+public static class StartupExtensions
+{
+    public static void AddDataInfrastructure(this IServiceCollection services)
+    {
+        services.AddSingleton<IConnectionResolver, DefaultConnectionResolver>();
+    }
+}
